refactor(tui): extract Update loop into update.go
internal/tui/tui.go の中核から Update(233 行)と、その唯一の呼び出し元である
resumeKeyboardProtocols を新規 internal/tui/update.go へ移動。同一 package 内の宣言移動のみで
ロジック・シグネチャ・挙動は不変(非テスト宣言 247 件が移動前と完全一致)。

最も改変頻度の高い Update を専用ファイルに隔離し、tui.go を構築+ライフサイクルの spine
(Run/normalizeOptions/newModel/Init/quit)に保つ。#331 のドメイン分割に、競合 PR #329 が
持っていた「Update 隔離」の利点を取り込む follow-up。
